[xlnx-versal-virt]: change DT compat string to match vxworks
VxWorks looks for "xlnx,versal" compatible string at the top level
node of the device tree to match the compatible string in the versal
board description found in the VxWorks Versal BSP. VxWorks does not
match the "xlnx-versal-virt" compatible string in the qemu generated
device tree for this machine model.

Change the compatible string in the device tree that qemu generates
for the xlnx-versal-virt board to match VxWorks BOARD_DESC for
xilinx versal.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>